requirements.txt
requests
confluent_kafka
kafka-python
numpy
pandas
datetime